backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
alembic==1.12.1
asyncpg==0.29.0
pandas==2.1.3
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0